[EasyActivity] Fix CI
